+++ namer/namer.go
@@ -37,10 +37,7 @@ func NewDefaultNamer(prefix string, hashNames []string, sigNames []string) Namer
 
 // GenerateNames all keys for an object name.
 func (n *DefaultNamer) GenerateNames(name string) ([]Key, error) {
-	switch {
-	case name == "":
-		return nil, errInvalidName(name, "should not be empty")
-	case strings.HasSuffix(name, "/"):
+	if strings.HasSuffix(name, "/") {
 		return nil, errInvalidName(name, "should not be prefix")
 	}