+++ operation/option.go
@@ -1,4 +1,6 @@
 package operation
 
 // Option contains configuration options for operations.
-type Option struct{}
+type Option struct {
+	WithPrefix bool
+}