+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - integrity.Validator: Ability to get ModRevision from a validated result (#23).
+- integrity.Typed: The possibility of prefix deletion (#24).
 
 ### Changed